Moved FifoQueue to separate file and minor cleanup. #7572
